+++ sdk/resourcemanager/containerservice/armcontainerservice/constants.go
@@ -7,7 +7,7 @@ package armcontainerservice
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
-	moduleVersion = "v7.0.0"
+	moduleVersion = "v7.1.0"
 )
 
 // AgentPoolMode - A cluster must have at least one 'System' Agent Pool at all times. For additional information on agent
@@ -15,6 +15,9 @@ const (
 type AgentPoolMode string
 
 const (
+	// AgentPoolModeGateway - Gateway agent pools are dedicated to providing static egress IPs to pods. For more details, see
+	// https://aka.ms/aks/static-egress-gateway.
+	AgentPoolModeGateway AgentPoolMode = "Gateway"
 	// AgentPoolModeSystem - System agent pools are primarily for hosting critical system pods such as CoreDNS and metrics-server.
 	// System agent pools osType must be Linux. System agent pools VM SKU must have at least 2vCPUs and 4GB of memory.
 	AgentPoolModeSystem AgentPoolMode = "System"
@@ -25,6 +28,7 @@ const (
 // PossibleAgentPoolModeValues returns the possible values for the AgentPoolMode const type.
 func PossibleAgentPoolModeValues() []AgentPoolMode {
 	return []AgentPoolMode{
+		AgentPoolModeGateway,
 		AgentPoolModeSystem,
 		AgentPoolModeUser,
 	}
@@ -38,6 +42,8 @@ const (
 	AgentPoolTypeAvailabilitySet AgentPoolType = "AvailabilitySet"
 	// AgentPoolTypeVirtualMachineScaleSets - Create an Agent Pool backed by a Virtual Machine Scale Set.
 	AgentPoolTypeVirtualMachineScaleSets AgentPoolType = "VirtualMachineScaleSets"
+	// AgentPoolTypeVirtualMachines - Create an Agent Pool backed by a Single Instance VM orchestration mode.
+	AgentPoolTypeVirtualMachines AgentPoolType = "VirtualMachines"
 )
 
 // PossibleAgentPoolTypeValues returns the possible values for the AgentPoolType const type.
@@ -45,6 +51,7 @@ func PossibleAgentPoolTypeValues() []AgentPoolType {
 	return []AgentPoolType{
 		AgentPoolTypeAvailabilitySet,
 		AgentPoolTypeVirtualMachineScaleSets,
+		AgentPoolTypeVirtualMachines,
 	}
 }
 
@@ -713,6 +720,30 @@ func PossibleOutboundTypeValues() []OutboundType {
 		OutboundTypeNone,
 		OutboundTypeUserAssignedNATGateway,
 		OutboundTypeUserDefinedRouting,
+	}
+}
+
+// PodIPAllocationMode - The IP allocation mode for pods in the agent pool. Must be used with podSubnetId. The default is
+// 'DynamicIndividual'.
+type PodIPAllocationMode string
+
+const (
+	// PodIPAllocationModeDynamicIndividual - Each node gets allocated with a non-contiguous list of IP addresses assignable to
+	// pods. This is better for maximizing a small to medium subnet of size /16 or smaller. The Azure CNI cluster with dynamic
+	// IP allocation defaults to this mode if the customer does not explicitly specify a podIPAllocationMode
+	PodIPAllocationModeDynamicIndividual PodIPAllocationMode = "DynamicIndividual"
+	// PodIPAllocationModeStaticBlock - Each node is statically allocated CIDR block(s) of size /28 = 16 IPs per block to satisfy
+	// the maxPods per node. Number of CIDR blocks >= (maxPods / 16). The block, rather than a single IP, counts against the Azure
+	// Vnet Private IP limit of 65K. Therefore block mode is suitable for running larger workloads with more than the current
+	// limit of 65K pods in a cluster. This mode is better suited to scale with larger subnets of /15 or bigger
+	PodIPAllocationModeStaticBlock PodIPAllocationMode = "StaticBlock"
+)
+
+// PossiblePodIPAllocationModeValues returns the possible values for the PodIPAllocationMode const type.
+func PossiblePodIPAllocationModeValues() []PodIPAllocationMode {
+	return []PodIPAllocationMode{
+		PodIPAllocationModeDynamicIndividual,
+		PodIPAllocationModeStaticBlock,
 	}
 }